added W2T2 string functions

+++ ConsoleAppAlfaExercises/arrayFuncs.cpp
@@ -8,7 +8,7 @@ unique_ptr<int[]> createArray(int a) {
 
 
 unique_ptr<int[]> fillArray(unique_ptr<int[]> a) {
-	for (size_t i = 0; i < sizeof(a) / sizeof(a[0]); i++)
+	for (int i = 0; i < sizeof(a) / sizeof(a[0]); i++)
 	{
 		a[i] = i;
 	}
